Changed BytesType to use 64 bit int by default.

+++ core/modules/core/types/bytes_type.php
@@ -2,7 +2,7 @@
 
 class BytesType extends \nmvc\AppType {
     public function getSQLType() {
-        return "int";
+        return "bigint";
     }
 
     public function getSQLValue() {
